chore: padronize with ios and send progresses as int

+++ packages/ads/android/src/main/kotlin/com/suamusica/smads/AdPlayerViewController.kt
@@ -124,8 +124,8 @@ class AdPlayerViewController(
             Timber.d("onAdEvent(%s)", adEvent)
         }
 
-        val duration = adPlayerManager?.adsDuration()?.toDouble()?.let { ceil(it).toLong() } ?: 0L
-        val position = adPlayerManager?.adsCurrentPosition()?.toDouble()?.let { ceil(it).toLong() } ?: 0L
+        val duration = adPlayerManager?.adsDuration()?.toInt()?.let { ceil(it.toDouble()).toInt() } ?: 0
+        val position = adPlayerManager?.adsCurrentPosition()?.toInt()?.let { ceil(it.toDouble()).toInt() } ?: 0
         Timber.d("onAdEvent(duration=%s, position=%s)", duration, position)
         callback.onAddEvent(AdEventOutput.fromAdEvent(adEvent, duration, position))
     }
